requirements.txt
pydantic
pytest
pyyaml
streamlit
